xerces-c: Do not build proto fuzzer for MSan

+++ projects/xerces-c/build.sh
@@ -10,22 +10,22 @@ make -j
 #cmake -DBUILD_SHARED_LIBS:BOOL=OFF ..
 
 cd $SRC
-rm -rf genfiles && mkdir genfiles && LPM/external.protobuf/bin/protoc xml.proto --cpp_out=genfiles
-cd libprotobuf-mutator
-#../LPM/external.protobuf/bin/protoc examples/xml/xml.proto --cpp_out=../genfiles
-cd $SRC
 
 $CXX $CXXFLAGS $LIB_FUZZING_ENGINE -std=c++11 \
         -Ixerces-c/src -Ixerces-c/src \
         xerces_fuzz_common.cpp parse_target.cpp -o $OUT/parse_target \
         xerces-c/src/.libs/libxerces-c.a
 
-$CXX $CXXFLAGS $LIB_FUZZING_ENGINE -std=c++11 \
-        -I xerces-c/src -Ixerces-c/build/src genfiles/xml.pb.cc xmlProtoConverter.cpp xerces_fuzz_common.cpp parse_target_proto.cpp \
-        -I libprotobuf-mutator/ \
-        -I genfiles \
-        -I LPM/external.protobuf/include \
-        -o $OUT/parse_target_proto xerces-c/src/.libs/libxerces-c.a \
-        LPM/src/libfuzzer/libprotobuf-mutator-libfuzzer.a \
-        LPM/src/libprotobuf-mutator.a \
-        LPM/external.protobuf/lib/libprotobuf.a
+if [[ $CFLAGS != *sanitize=memory* ]]; then
+	rm -rf genfiles && mkdir genfiles && LPM/external.protobuf/bin/protoc xml.proto --cpp_out=genfiles
+
+	$CXX $CXXFLAGS $LIB_FUZZING_ENGINE -std=c++11 \
+	        -I xerces-c/src -Ixerces-c/build/src genfiles/xml.pb.cc xmlProtoConverter.cpp xerces_fuzz_common.cpp parse_target_proto.cpp \
+	        -I libprotobuf-mutator/ \
+	        -I genfiles \
+	        -I LPM/external.protobuf/include \
+	        -o $OUT/parse_target_proto xerces-c/src/.libs/libxerces-c.a \
+	        LPM/src/libfuzzer/libprotobuf-mutator-libfuzzer.a \
+	        LPM/src/libprotobuf-mutator.a \
+	        LPM/external.protobuf/lib/libprotobuf.a
+fi
